Move `publish` action to `PublicationController`
and rename to `create` to match our other controller actions.
